test: Rewrite tests to use Spring Boot + devtools for reload (#13020) (CP: 23.0)
Jetty 10 does not support using scan combined with jetty:start.
This replaces tests need redeployment during the test with
Spring Boot using Spring Boot devtools
