+++ flow-tests/test-redeployment/src/main/java/com/vaadin/flow/uitest/ui/SessionValueView.java
@@ -7,21 +7,20 @@ import com.vaadin.flow.component.html.NativeButton;
 import com.vaadin.flow.router.Route;
 import com.vaadin.flow.server.VaadinSession;
 import com.vaadin.flow.server.WrappedSession;
-import com.vaadin.flow.uitest.servlet.ViewTestLayout;
 
-@Route(value = "com.vaadin.flow.uitest.ui.SessionValueView", layout = ViewTestLayout.class)
+@Route(value = "com.vaadin.flow.uitest.ui.SessionValueView")
 public class SessionValueView extends Div {
 
     // This ensures the view is not serializable
     private Object preventSerialization = new Object();
 
+    private UUID viewId = UUID.randomUUID();
+
     // Test that session values outside VaadinSession are preserved on reload
     public SessionValueView() {
 
         final NativeButton triggerButton = new NativeButton("Trigger reload",
-                event -> {
-                    WebpackDevServerPortView.triggerJettyReload();
-                });
+                event -> Application.triggerReload());
         triggerButton.setId(WebpackDevServerPortView.TRIGGER_RELOAD_ID);
         add(triggerButton);
 
@@ -35,6 +34,10 @@ public class SessionValueView extends Div {
         Div div = new Div();
         div.setId("customAttribute");
         div.setText("The custom value in the session is: " + customAttribute);
+        add(div);
+
+        div = new Div();
+        div.setText("The view id is: " + viewId.toString());
         add(div);
 
     }
